+++ tests/testthat/test-mysql.R
@@ -4,9 +4,9 @@ skip_on_cran()
 
 db <- dbxConnect(adapter="rmysql", dbname="dbx_test")
 
-DBI::dbExecute(db, "DROP TABLE IF EXISTS events")
+dbxExecute(db, "DROP TABLE IF EXISTS events")
 json_type <- if (Sys.getenv("TRAVIS") == "") "JSON" else "TEXT"
-DBI::dbExecute(db, paste0("CREATE TABLE events (id INT NOT NULL AUTO_INCREMENT PRIMARY KEY, city VARCHAR(255), counter INT, speed FLOAT, distance DECIMAL(5, 2), created_on DATE, updated_at DATETIME(6), deleted_at TIMESTAMP(6) NULL DEFAULT NULL, open_time TIME, active BOOLEAN, properties ", json_type, ", image BLOB)"))
+dbxExecute(db, paste0("CREATE TABLE events (id INT NOT NULL AUTO_INCREMENT PRIMARY KEY, city VARCHAR(255), counter INT, speed FLOAT, distance DECIMAL(5, 2), created_on DATE, updated_at DATETIME(6), deleted_at TIMESTAMP(6) NULL DEFAULT NULL, open_time TIME, active BOOLEAN, properties ", json_type, ", image BLOB)"))
 
 runTests(db)