+++ tests/testthat/test-postgres.R
@@ -4,8 +4,8 @@ skip_on_cran()
 
 db <- dbxConnect(adapter="rpostgres", dbname="dbx_test")
 
-DBI::dbExecute(db, "DROP TABLE IF EXISTS events")
-DBI::dbExecute(db, "CREATE TABLE events (id SERIAL PRIMARY KEY, city TEXT, counter INT, speed FLOAT, distance DECIMAL, created_on DATE, updated_at TIMESTAMP, deleted_at TIMESTAMPTZ, open_time TIME, close_time TIMETZ, active BOOLEAN, properties JSON, propertiesb JSONB, image BYTEA)")
+dbxExecute(db, "DROP TABLE IF EXISTS events")
+dbxExecute(db, "CREATE TABLE events (id SERIAL PRIMARY KEY, city TEXT, counter INT, speed FLOAT, distance DECIMAL, created_on DATE, updated_at TIMESTAMP, deleted_at TIMESTAMPTZ, open_time TIME, close_time TIMETZ, active BOOLEAN, properties JSON, propertiesb JSONB, image BYTEA)")
 
 runTests(db)