+++ tests/testthat/test-redshift.R
@@ -4,7 +4,7 @@ skip("redshift")
 
 db <- dbxConnect(url="rpostgresql://...")
 
-DBI::dbExecute(db, "DROP TABLE IF EXISTS events")
-DBI::dbExecute(db, "CREATE TABLE events (id INT IDENTITY(0, 1), city TEXT, counter INT, speed FLOAT, distance DECIMAL, created_on DATE, updated_at TIMESTAMP, deleted_at TIMESTAMPTZ, open_time TEXT, close_time TEXT, active BOOLEAN, properties TEXT, propertiesb TEXT, image TEXT, PRIMARY KEY(id))")
+dbxExecute(db, "DROP TABLE IF EXISTS events")
+dbxExecute(db, "CREATE TABLE events (id INT IDENTITY(0, 1), city TEXT, counter INT, speed FLOAT, distance DECIMAL, created_on DATE, updated_at TIMESTAMP, deleted_at TIMESTAMPTZ, open_time TEXT, close_time TEXT, active BOOLEAN, properties TEXT, propertiesb TEXT, image TEXT, PRIMARY KEY(id))")
 
 runTests(db, redshift=TRUE)