+++ tests/testthat/test-sqlite.R
@@ -2,7 +2,7 @@ context("sqlite")
 
 db <- dbxConnect(adapter="rsqlite", dbname=":memory:")
 
-DBI::dbExecute(db, "CREATE TABLE events (id INTEGER PRIMARY KEY AUTOINCREMENT, city VARCHAR(255), counter INT, speed FLOAT, distance DECIMAL, created_on DATE, updated_at DATETIME, open_time VARCHAR(255), active BOOLEAN, properties TEXT, image BLOB)")
+dbxExecute(db, "CREATE TABLE events (id INTEGER PRIMARY KEY AUTOINCREMENT, city VARCHAR(255), counter INT, speed FLOAT, distance DECIMAL, created_on DATE, updated_at DATETIME, open_time VARCHAR(255), active BOOLEAN, properties TEXT, image BLOB)")
 
 runTests(db)